resume_mapping/requirements.txt
boto3 
python-dotenv
streamlit
faiss-cpu
PyPDF2
scikit-learn
asyncio